requirements.txt
flask==2.0.1
firebase-admin==5.0.3
python-dotenv==0.19.0
flask-login==0.5.0
werkzeug==2.0.1 